created and test DeviceTimeNoRTC - the time sync test wasn't actually running lol also it was failing so i fixed that
